Todo Example of DI from DAL to SL
