+++ Uroboro/Uroboro.SL.WebAPI/Startup.cs
@@ -7,6 +7,8 @@ using Microsoft.Extensions.Hosting;
 using Microsoft.IdentityModel.Tokens;
 using Microsoft.OpenApi.Models;
 using System.Text;
+using Uroboro.BL.Managers;
+using Uroboro.DAL.InMemory.Repos;
 using Uroboro.SL.WebAPI.Services;
 
 namespace Uroboro.SL.WebAPI
@@ -54,6 +56,9 @@ namespace Uroboro.SL.WebAPI
                         builder.AllowAnyMethod().AllowAnyHeader().AllowAnyOrigin().Build();
                     });
             });
+
+            // Add controller dependencies
+            services.AddTodoManager();
 
             services.AddControllers();
             services.AddSwaggerGen(c =>
